Moved register defines to cpp. Changed to const static.
Const static allows compiler to check for type conflicts.

+++ src/heartRate.h
@@ -61,7 +61,7 @@
  #include "WProgram.h"
 #endif
 
-bool checkForBeat(long sample);
-int16_t averageDCEstimator(long *p, uint16_t x);
+bool checkForBeat(int32_t sample);
+int16_t averageDCEstimator(int32_t *p, uint16_t x);
 int16_t lowPassFIRFilter(int16_t din);
 int32_t mul16(int16_t x, int16_t y);
